solve worning in events

+++ include/ara/com/proxy_skeleton/skeleton/event.hpp
@@ -163,7 +163,7 @@ namespace ara
                         bool op = false;
                         std::vector<uint8_t> _data = sd_msg.GetPayload();
                         SOMEIP_MESSAGE::Message R_msg(
-                            SOMEIP_MESSAGE::Message_ID{ m_service->m_service_id.GetInstanceId(), sd_msg.MessageId().method_id|0x8000},
+                            SOMEIP_MESSAGE::Message_ID{ (uint16_t)m_service->m_service_id.GetInstanceId(),(uint16_t) (sd_msg.MessageId().method_id|0x8000)},
                             SOMEIP_MESSAGE::Request_ID{5,6},
                             2, // protocol version
                             7, // Interface Version
